initial set up for exec check
